+++ CHANGELOG.md
@@ -358,3 +358,4 @@ with new `PluginSerializer`
 ### 1.3.1
 
 * Update `TelegramBotAPI` version `0.14.0` -> `0.14.1`
+* In `Launch` added `flowFilter` and updates collecting rewritten on flows